change data type to decimal for debit and credit journal item

+++ Com.DanLiris.Service.Purchasing.Lib/Enums/COAGenerator.cs
@@ -338,8 +338,8 @@ namespace Com.DanLiris.Service.Purchasing.Lib.Enums
     {
         public COA COA { get; set; }
         public string Remark { get; set; }
-        public double? Debit { get; set; }
-        public double? Credit { get; set; }
+        public decimal? Debit { get; set; }
+        public decimal? Credit { get; set; }
     }
 
     public class COA
